Exclude AutoModerator from a drawing.

+++ backend/app/cache/cache.py
@@ -203,6 +203,8 @@ class GogCache:
             db.session.flush()
 
     def filter_comment(self, comment):
+        if comment.author and self.reddit.is_user_special(comment.author.name):
+            return
         logging.info('Processing comment {}. Thread: {}.'.format(comment.id, comment.submission.url))
         if self.reddit.is_deleted(comment):
             self.delete_comment_from_db(comment)
